refactor(playground): extract usePlaygroundPrintDialog
